README and a little analysis
Copied over the README from the Introduction information in the Sheets. Tried to do a little analysis coding for farming levels, and I'm think Excel/Sheets will be the best tools for this analysis in the end. R is great, but I don't need the complexity (at least not yet).